Removed get_borders and fixed resulting mess, added get_slices_multigrid
