+++ src/porespy/generators/__init__.py
@@ -44,6 +44,5 @@ from ._imgen import *
 from ._pseudo_packings import *
 from ._noise import *
 from ._spheres_from_coords import *
-from ._borders import *
 from ._fractals import *
 from ._micromodels import *
