+++ src/porespy/generators/_pseudo_packings.py
@@ -8,7 +8,6 @@ from skimage.morphology import ball, disk
 from porespy.filters import trim_disconnected_blobs
 from porespy.tools import (
     _insert_disk_at_point,
-    get_border,
     ps_round,
     unpad,
     get_edt,
@@ -51,7 +50,7 @@ def _random_spheres2(
     machinery as the pseudo packing generators.  It is not as fast as the original
     though.
     """
-
+    from porespy.generators import borders
     if seed is not None:
         _set_seed(seed)  # Initialize rng so numba sees it
         np.random.seed(seed)  # Also initialize numpys rng
@@ -74,7 +73,7 @@ def _random_spheres2(
 
     # Deal with edge mode
     if edges == 'contained':
-        border = get_border(im.shape, thickness=1, mode='faces')
+        border = borders(im.shape, thickness=1, mode='faces')
         mask[border] = False
 
     # Generate mask of valid insertion points
@@ -345,6 +344,7 @@ def pseudo_electrostatic_packing(
     to view online example.
 
     """
+    from porespy.generators import borders
     if seed is not None:  # Initialize rng so numba sees it
         _set_seed(seed)
         np.random.seed(seed)
@@ -365,15 +365,15 @@ def pseudo_electrostatic_packing(
         mask = dt >= abs(protrusion)
 
     if edges == 'contained':
-        borders = get_border(mask.shape, thickness=1, mode='faces')
-        mask[borders] = 0
+        border = borders(mask.shape, thickness=1, mode='faces')
+        mask[border] = 0
 
     if sites is None:
         dt = edt(mask)
         dt = spim.gaussian_filter(dt, sigma=0.5)
         strel = ps_round(r, ndim=im.ndim, smooth=True)
         sites = (spim.maximum_filter(dt, footprint=strel) == dt)*(mask > 0)
-        if np.any(dt == np.inf) or np.all(dt == dt[0]):  # In case above method failed.
+        if np.any(dt == np.inf) or np.all(dt == dt[0]):  # If above method failed
             sites = np.zeros_like(im)
             inds = tuple((np.array(im.shape)/2).astype(int))
             sites[inds] = True
